Updating the mapping extensions
Updating the mapping extensions to enable multiple rig/skeleton and expression set maps in an instance of the extension (given it's a top-level extension).

There were also some contradictions in these; which I cleaned up and tried to address.

These will be touched up once we make the switch to the KHR_character base